good gob

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int i;
 long int li;
 long long int lli;
 float f;
-printf("Size of a char: %lu byte(s),\n", sizeof(c));
-printf("Size of an int: %lu byte(s),\n", sizeof(i));
-printf("size of a long int: %lu byte(s),\n", sizeof(li));
-printf("Size of a long long int: %lu byte(s),\n", sizeof(lli));
-printf("size of a float: %lu byte(s),\n", sizeof(f));
+printf("Size of a char: %lu byte(s),\n", (unsigned long)sizeof(c));
+printf("Size of an int: %lu byte(s),\n", (unsigned long)sizeof(i));
+printf("size of a long int: %lu byte(s),\n", (unsigned long)sizeof(li));
+printf("Size of a long long int: %lu byte(s),\n", (unsigned long)sizeof(lli));
+printf("size of a float: %lu byte(s),\n", (unsigned long)sizeof(f));
 return (0);
 }
